Feat. Nearly Functional Translator Implementation. Missing Initial State Conditions.
Is currently able to be a little smart about which actions it needs to generate, identifies
predicates which will never change, and only instantiates those actions which are present in the initial condition.

Does not identify, or handle explicit labelling of, predicates to variables.

The test cases are also rather basic and need some extension.