Feat.: First functioning Translator Prototype.
Had to apply some hacky changes to formula and symbols for making asp code safe, but the translator does minimally function.

Have successfully used it to solve the acrobatics/01 domain. Supports oneof clauses if they are at the top (in the standard normal form).

Needs some refactoring and polish, and perhaps some more rigorous tests, but the functional prototype exists.