Fix: No longer instantiate an excessive number of actions.
There was an issue detecting actions with unchanging preconditions, the required preconditions were not being instantiated correctly.
This lead to an excessive number of grounded actions being generated, for various parameter combinations which were not possible.

This is now fixed. Actions and grounded actions are also hashable, in case some actions were generated as duplicates.